server: handle stop in a separate thread
Fix: WAZO-3131

why: bullseye cheroot version (8.5.2) is now waiting for the main
serving thread to complete its current tick before stopping. If we
execute stop in the same thread that main thread, we will have a dead
lock.

Moreover, signal handler should not execute a lot of code (or wait) and
only set some variable/flag. So it's not that bad to use threading at
this level.